Fix WP test to start from clean slate
